+++ src/components/Multiselect/Common/index.ts
@@ -1,3 +1,3 @@
-export * from "./useMultiselect";
-export * from "./MultiselectWrapper";
 export * from "./Multiselect.css";
+export * from "./MultiselectWrapper";
+export * from "./useMultiselect";
